+++ zarrs/src/array/element.rs
@@ -3,7 +3,9 @@ use std::mem::ManuallyDrop;
 use itertools::Itertools;
 use ArrayError::IncompatibleElementType as IET;
 
-use super::{convert_from_bytes_slice, transmute_to_bytes, ArrayBytes, ArrayError, DataType};
+use super::{
+    convert_from_bytes_slice, transmute_to_bytes, ArrayBytes, ArrayError, DataType, RawBytesOffsets,
+};
 
 /// A trait representing an array element type.
 pub trait Element: Sized + Clone {
@@ -184,13 +186,21 @@ macro_rules! impl_element_string {
                     len = len.checked_add(element.len()).unwrap();
                 }
                 offsets.push(len);
+                let offsets = unsafe {
+                    // SAFETY: The offsets are monotonically increasing.
+                    RawBytesOffsets::new_unchecked(offsets)
+                };
 
                 // Concatenate bytes
                 let mut bytes = Vec::with_capacity(usize::try_from(len).unwrap());
                 for element in elements {
                     bytes.extend_from_slice(element.as_bytes());
                 }
-                Ok(ArrayBytes::new_vlen(bytes, offsets))
+                let array_bytes = unsafe {
+                    // SAFETY: The last offset is the length of the bytes.
+                    ArrayBytes::new_vlen_unchecked(bytes, offsets)
+                };
+                Ok(array_bytes)
             }
         }
     };
@@ -238,11 +248,19 @@ macro_rules! impl_element_binary {
                     len = len.checked_add(element.len()).unwrap();
                 }
                 offsets.push(len);
+                let offsets = unsafe {
+                    // SAFETY: The offsets are monotonically increasing.
+                    RawBytesOffsets::new_unchecked(offsets)
+                };
 
                 // Concatenate bytes
                 let bytes = elements.concat();
 
-                Ok(ArrayBytes::new_vlen(bytes, offsets))
+                let array_bytes = unsafe {
+                    // SAFETY: The last offset is the length of the bytes.
+                    ArrayBytes::new_vlen_unchecked(bytes, offsets)
+                };
+                Ok(array_bytes)
             }
         }
     };